data-preprocessor/requirements.txt
paho-mqtt
pandas
pika
